8/23

+++ leet_code/generate.py
@@ -1,10 +1,12 @@
 
 import os
 
-files = os.listdir('./python')
+root = os.path.dirname(__file__)
+files = os.listdir(os.path.join(root,'python'))
 
 for file in files:
     file = file[:-2]+'md'
-    file = os.path.join('./solutions',file)
-    f = open(file, 'w')
-    f.close()
+    file = os.path.join(root, 'solutions',file)
+    if not os.path.exists(file):
+        f = open(file, 'w')
+        f.close()
